Move BlackRendererSettings to __init__.py
Otherwise `flake8` is too stupid to find it even though it works fine in
actual CPython interpreter.
